+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 
 ## Current Development
 
+
+- [NAL-1347](https://jira.skatelescope.org/browse/NAL-1347)
+    - [Added] Added a PVDataProduct, PVIndex and PVInterface class, refactoring the PV interface in preparation to ease the integration with the DLM database.
+
 - [NAL-1323](https://jira.skatelescope.org/browse/NAL-1323)
     - [Changed] Added the functionality to allow annotations to be updated.
     - [Changed] Added a check to return nothing if PostgresSQL is not available when attempting to use annotation endpoints.